--- frontend/src/components/activity-feed/hooks/activity.d.ts
@@ -1,11 +1,0 @@
-import { IActivityFeedItem } from 'interfaces/IActivityFeedItem';
-
-export type ActivityFeedById = {
-  [uuid: string]: IActivityFeedItem;
-};
-
-export interface IEventActivity {
-  type: string;
-  created_at: string;
-  args?: any;
-}
