+++ frontend/src/components/activity-feed/interfaces/IActivityFeedById.ts
@@ -1,0 +1,5 @@
+import { IActivityFeedItem } from 'interfaces/IActivityFeedItem';
+
+export interface IActivityFeedById {
+  [uuid: string]: IActivityFeedItem;
+}
