+++ frontend/src/utils/contract.ts
@@ -1,0 +1,5 @@
+// the last 8 characters of the deployed job contract
+export const getSmartContractId = () => {
+  const address = String(process.env.REACT_APP_JOB_CONTRACT_ADDRESS);
+  return address.substring(address.length - 8).toLowerCase();
+};
